Usando Objetos nas Entidades

+++ src/main/java/com/abutua/productbackend/models/Product.java
@@ -15,7 +15,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String description;
@@ -25,10 +25,10 @@ public class Product {
 
     private boolean promotion;
     private boolean newProduct;
-    private double price;
+    private Double price;
 
     // Métodos Construtores
-    public Product(int id, String name, String description, Category category, boolean promotion, boolean newProduct,
+    public Product(Long id, String name, String description, Category category, boolean promotion, boolean newProduct,
             double price) {
         this.id = id;
         this.name = name;
@@ -39,7 +39,7 @@ public class Product {
         this.price = price;
     }
 
-    public Product(int id, String name, double price) {
+    public Product(Long id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -50,11 +50,11 @@ public class Product {
     }
 
     // Métodos
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
