+++ src/main/java/com/abutua/productbackend/services/ProductService.java
@@ -22,7 +22,7 @@ public class ProductService {
     @Autowired
     private CategoryService categoryService;
     
-    public Product getById(int id) {
+    public Product getById(long id) {
         Product product = productRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Product not found"));
         
@@ -37,12 +37,12 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public void deleteById(int id) {
+    public void deleteById(long id) {
         Product product = getById(id);
         productRepository.delete(product);
     }
 
-    public void update(@PathVariable int id, @RequestBody Product productUpdate) {
+    public void update(@PathVariable long id, @RequestBody Product productUpdate) {
         Product product = getById(id);
 
         if (productUpdate.getCategory() == null) {
